tests/wpt/web-platform-tests/speculation-rules/prerender/about-blank-iframes.html

<!doctype html>
<!--
Tests for about:blank iframes `document.prerendering` state
and prerenderingchange event.
-->
<title>Test about:blank iframes prerendering state</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/utils.js"></script>
<body>
<script>

setup(() => assertSpeculationRulesIsSupported());

promise_test(async t => {
  const channel = new BroadcastChannel('test-channel');

  const gotMessage = new Promise(resolve => {
    channel.addEventListener('message', e => {
      resolve(e.data);
    }, {once: true});
  });

  // Make the window to start the prerender.
  const url = `resources/about-blank-iframes.html`;
  window.open(url, '_blank', 'noopener');

  const msg = await gotMessage;
  assert_equals(msg, 'PASS');
}, 'about:blank iframes');
</script>
</body>
